backend/requirements.txt
addict==2.4.0
fiftyone
gradio
huggingface_hub
ipdb==0.13.13
matplotlib
numpy>=1.26.0
opencv_python_headless==4.10.0.84
Pillow==11.1.0
pycocotools==2.0.8
PyYAML==6.0.2
rembg==2.0.59
Requests==2.32.3
setuptools==75.8.0
supervision==0.25.1
termcolor==2.5.0
timm==1.0.13
torch==2.5.1
torchvision==0.20.1
tqdm==4.66.5
transformers==4.48.0
typer==0.15.1
yapf==0.43.0
